Added forcefield directory for UFF and created atom type assignement file specific to that forcefield.
