+++ lammps_tools/forcefields/uff/parameters/uff_nonbonded.py
